backend/requirements.txt
Flask==3.1.1
requests==2.32.4
python-dotenv==1.0.1
gunicorn==23.0.0

